requirements.txt
atlassian-python-api==3.41.18
markitdown==0.0.1a3
beautifulsoup4==4.12.3
pandas==2.2.3
